requirements.txt
flask
deepface
opencv-python
requests
matplotlib
tf-keras
minio
python-dotenv